rename: 같은 파일 충돌 대비하여 /apis/userApi.js 로 파일이름 변경 (#113)
